% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absorbance_to_beam_c.R
\name{absorbance_to_beam_c}
\alias{absorbance_to_beam_c}
\title{Convert absorbance measurements to beam attenuation}
\usage{
absorbance_to_beam_c(
  ANGLE,
  TURB,
  L,
  TRANSMITTANCE,
  UNITS = "JTU",
  WAVELENGTH = 800
)
}
\arguments{
\item{ANGLE}{Acceptance angle of the spectrophotometer}

\item{TURB}{Turbidity measurement in units of JTU, NTU}

\item{L}{Length of spectrophotometer cuvette}

\item{TRANSMITTANCE}{Percentage of transmitted light reaching sensor}

\item{UNITS}{Turbidity measurement units}

\item{WAVELENGTH}{Transmitted wavelength for spectrophotometer}
}
\description{
Function converts absorbance measurements obtained using spectophotometer to beam attenuation based on the acceptance angle of the spectrophotometer, turbidity standard, measurement units for turbidity, cuvette length, and transmittance wavelength.
}
