% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICc_npar.R
\name{AICc_npar}
\alias{AICc_npar}
\title{Calculate AICc using a user-specified number of parameters}
\usage{
AICc_npar(object, npar = NA, npar.out = FALSE)
}
\arguments{
\item{object}{mle2 model object}

\item{npar}{Number of parameters in the model. If NA, will use the length of fullcoef(object)}

\item{npar.out}{Returns number of parameters and AICc.}

\item{use.AIC}{Use AIC instead of AICc?}
}
\description{
This function is used to estimate AICc in cases where it is necessary to pass estimated fixed values of parameters into a model in order to estimate other parameters. Arises when using step functions with a discontinuous likelihood profile.
}
\author{
S.K. Rohan \email{skrohan@uw.edu}
}
