% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_chars_for_plots.R
\name{convert_chars_for_plots}
\alias{convert_chars_for_plots}
\title{Convert model parameters to expressions for plotting}
\usage{
convert_chars_for_plots(x)
}
\arguments{
\item{x}{Data frame with a variable column that includes parameters to be converted to expressions}
}
\value{
Data frame with character string that can be converted to an expression.
}
\description{
Function used in Rohan et al. GVRDM paper
}
