% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\title{Diagnostic plots for the AUM and GVRDM}
\usage{
diagnostic_plots(
  rr = NA,
  out = NA,
  cc = NA,
  Eb = NA,
  sigma = NA,
  cont_shape = NA,
  NVrd = NA,
  NVrr = NA,
  NVsigma = NA
)
}
\arguments{
\item{rr}{Observed reaction distances for the visual component}

\item{out}{Fitted reaction distance for the visual component}

\item{cc}{Effective attenuation coefficient or beam attenuation coefficient}

\item{Eb}{light}

\item{sigma}{standard deviation for the visual component}

\item{cont_shape}{Optional. Manually-specified dynamic scaling function values.}

\item{NVrd}{Observed non-visual reaction distances for the non-visual component}

\item{NVrr}{Fitted non-visual reaction distances for the non-visual component}

\item{NVsigma}{Standard deviation for non-visual reaction distance}
}
\value{
Returns a list containing diagnostic plots.
}
\description{
Function called by fit_gvrdm to generate model diagnostic plots.
}
