% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_scaling.R
\name{dynamic_scaling}
\alias{dynamic_scaling}
\title{Dynamic scaling function}
\usage{
dynamic_scaling(beta, delta, kk, cc)
}
\arguments{
\item{beta}{Dynamic scaling function beta}

\item{delta}{Dynamic scaling function delta}

\item{kk}{Dynamic scaling function h}

\item{cc}{Effective attenuation coefficient or beam attenuation coefficient}

\item{return}{A vector of dynamic scaling function values.}
}
\description{
Function to plot a dynamic scaling function. Parameters correspond with model parameters.
}
