% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_lambertW.R
\name{f.lambertW}
\alias{f.lambertW}
\title{Lambert W function}
\usage{
f.lambertW(x, tolerance = 1e-10, maxit = 50)
}
\arguments{
\item{x}{Value}

\item{tolerance}{Acceptable precision}

\item{maxit}{Maximum number of iterations to find the answer}
}
\description{
Implementation of the Lambert W function
}
