% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_eval.R
\name{fit_eval}
\alias{fit_eval}
\title{Function to fit models}
\usage{
fit_eval(
  dat,
  mod,
  Ke = NA,
  alpha = NA,
  sigma = NA,
  NVthreshold = NA,
  silent = F
)
}
\arguments{
\item{mod}{mle2 model object.}

\item{Ke}{Optional. Default (NA) results in searching for the parameter in the model object.}

\item{alpha}{Optional. Default (NA) results in searching for the parameter in the model object.}

\item{sigma}{Optional. Default (NA) results in searching for the parameter in the model object.}

\item{NVthreshold}{Optional. Default (NA) results in searching for the parameter in the model object.}

\item{silent}{Should diagnostic plots be generated? Passed to diagnostic_plots function.}

\item{data}{Data set as a list, passed in the same way as for an mle2 model.}

\item{tt}{Optional. Default (NA) results in searching for the parameter in the model object.}
}
\value{
A list with the input, model fits, R-squared, and root mean square log error
}
\description{
Uses model and predictors to generate fit. Also returns root means squared log error and R-squared.
}
