% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naka_rushton.R
\name{naka_rushton}
\alias{naka_rushton}
\title{Modified Naka-Rushton function}
\usage{
naka_rushton(Rmax = 1, Eb, Ke, nn = 1, ss = 1, BB = 0)
}
\arguments{
\item{Rmax}{Maximum retinal response (default = 1)}

\item{Eb}{Incident light}

\item{Ke}{Half-saturation parameter}

\item{nn}{Shape parameter (Peirce 2007)}

\item{ss}{Shape parameter}

\item{BB}{Baseline signal without any incident light.}
}
\value{
Returns a vector of values of the Naka-Rushton function.
}
\description{
Contrast function (Naka-Rushton) with the flexibility to fit additional shapes, as described in Peirce (2007). Either of the function arguments nn or ss can be considered the equivalent of the alpha parameter.
}
\references{
Peirce, J.W., 2007. The potential importance of saturating and supersaturating contrast response functions in visual cortex. J. Vis. 7, 1–10. https://doi.org/10.1167/7.6.13
}
