% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLL_dynamic_contrast.R
\name{NLL_dynamic_contrast}
\alias{NLL_dynamic_contrast}
\title{Function for fitting the GVRDM and Aksnes and Utne model}
\usage{
NLL_dynamic_contrast(
  rr,
  cc,
  Ap = NA,
  C0 = NA,
  Eprime = NA,
  prey.types = NA,
  prey = NA,
  prey.size = NA,
  tt = NA,
  tt1 = NA,
  tt2 = NA,
  tt3 = NA,
  tt4 = NA,
  Ke,
  Eb,
  angle = NA,
  kd = NA,
  kd.mult = NA,
  beta = NA,
  kk = NA,
  delta = NA,
  alpha = 1,
  sigma = NA,
  rr.log = TRUE,
  NVrd = NA,
  NVthreshold = 0,
  NVsigma = NA,
  fit.model = TRUE,
  fit.obs = TRUE,
  ccoffset = NA,
  silent = FALSE,
  ...
)
}
\arguments{
\item{rr}{Reaction distance}

\item{cc}{Effective attenuation coefficient or beam attenuation coefficient}

\item{Ap}{Prey area. If Ap and C0 are not provided, tt will be estimated.}

\item{C0}{Prey inherent contrast. If Ap and C0 are not provided, tt will be estimated.}

\item{Eprime}{Composite saturation parameter.}

\item{prey.types}{A vector specifying unique prey name categories that are passed to 'prey.' Used if parameters are simultaneously estimated for multiple prey types.}

\item{prey}{Vector of prey names for each observation.}

\item{prey.size}{Vector of prey size}

\item{tt}{T parameter for prey type 0. Default = NA}

\item{tt1}{T parameter for prey type 1. Default = NA}

\item{tt2}{T parameter for prey type 2. Default = NA}

\item{tt3}{T parameter for prey type 3. Default = NA}

\item{tt4}{T parameter for prey type 4. Default = NA}

\item{Ke}{Half-saturation constant}

\item{Eb}{# Light level}

\item{angle}{Nadir viewing angle in degrees}

\item{kd}{Diffuse attenuation coefficient of downwelling irradiance}

\item{kd.mult}{Multiplier to convert beam attenuation}

\item{beta}{Dynamic scaling function intercept parameter}

\item{delta}{Dynamic scaling function shape parameter}

\item{alpha}{Naka-Rushton exponent. Default = 1}

\item{sigma}{Standard deviation of visual reaction distance}

\item{rr.log}{Are visual errors lognormal distribution? Default = TRUE}

\item{NVrd}{Non-visual reaction distance. Default = NA (only fit visual component of the model)}

\item{NVthreshold}{Light threshold for non-visual reaction. Default = NA(only fit visual component of the model)}

\item{NVsigma}{Standard deviation of non-visual reaction distance. Default = NA(only fit visual component of the model)}

\item{fit.model}{Logical.Should the model be fitted? If FALSE, produces model diagnostics and predictions based on provided parameters.}

\item{fit.obs}{Logical. Should observations be used to calculate a likelihood?}

\item{ccoffset}{Shift kappa or beam attenuation? Default = NA results in no shift.}

\item{silent}{Logical. Should diagnostic plots be produced?}

\item{hh}{Dynamic scaling function rate parameter}

\item{return}{If fit.model is TRUE, returns the negative log likelihood. If fit is FALSE, returns model diagnostics and predictions using parameters that are passed to the model.}
}
\description{
Generalized visual reaction distance model and Aksnes and Utne model. Identical to fit_gvrdm. Included for backward compatibility.
}
