% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_rd.R
\name{rmse_rd}
\alias{rmse_rd}
\title{Root mean square error lor root mean square log error}
\usage{
rmse_rd(obs, fit, log.trans = FALSE)
}
\arguments{
\item{obs}{Vector of observed values}

\item{fit}{Vector of fitted values}

\item{log.trans}{Should observed and fitted vectors be log-transformed to calculate RMSLE?}
}
\value{
Vector of RMSE or RMSLE.
}
\description{
Function to calculate the root mean square error (RMSE) or root mean square log error (RMSLE).
}
